New windows_task attribute to control command escaping
Old windows_task implementation was passing the command to schtasks:
In order to work fine, it enclosed the command in double quote, meaning
that you would have to escape the double quotes of your command to have
schtasks interprete it correctly.

New advanced_task LWRP directly doesn't have to enclose the command
with double quotes, so the extra escaping is useless.
In order to be backward compatible the windows_task LWRP automaticaly
strip the backslashes, but this behavior can be disabled with the new
windows_task attribute `escape_command`.