Correct http_acl resource updates when using only user (no sddl)
Before this patch, http_acl resource was not able to update acl when
user property was changed.
This was due to implementation stopping at sddl comparison when sddl
line was present on the result of showacl command (which always seems to be the
case on our servers)

This patch attempt to fix this to allow change of user to be taken into
account.

Change-Id: I9a894174f439ba201feae53d545f44669f06a20a
JIRA: WTH-1395